pyproject.toml
[tool.poetry]
name = "llamate"
version = "0.1.0"
description = "Simple model management for llama-swap"
authors = ["Llamate Contributors"]
readme = "README.md"
license = "MIT"
packages = [{include = "llamate"}]
keywords = ["llama-cpp", "llama-server", "model-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
huggingface-hub = "^0.21.0"
click = "^8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^4.0"
black = "^24.0"
mypy = "^1.8"

[tool.poetry.scripts]
llamate = "llamate.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"